Add files via upload
This is the procedure I made for the failure of the air-conditioning chiller RP-1043. The sequence may be messy. It includes the extended model of LSTM, including BiLSTM, GRU, NLSTM and other methods. The content also includes the principal component of the data to reduce the dimensionality for visualization, and the sliding window method to adjust the input form of the LSTM and so on.